+++ kart/sqlalchemy/create_engine.py
@@ -30,6 +30,22 @@ def _pool_class():
     return NullPool if "PYTEST_CURRENT_TEST" in os.environ else None
 
 
+def sqlite_engine(path):
+    """
+    An engine for non-spatial, non-GPKG sqlite databases.
+    """
+
+    def _on_connect(pysqlite_conn, connection_record):
+        pysqlite_conn.isolation_level = None
+        dbcur = pysqlite_conn.cursor()
+        dbcur.execute("PRAGMA journal_mode = 'wal';")
+        dbcur.execute("PRAGMA foreign_keys = ON;")
+
+    engine = sqlalchemy.create_engine(f"sqlite:///{path}", module=sqlite)
+    sqlalchemy.event.listen(engine, "connect", _on_connect)
+    return engine
+
+
 def gpkg_engine(path):
     def _on_connect(pysqlite_conn, connection_record):
         pysqlite_conn.isolation_level = None